[aivot] add orientation support
Orientation ensures robot's x-axis is aligned with a vector originating at current robot position and ending at frontier centroid. In addition to orientation, the angular distance is calculated as a simple absolute difference from current robot orientation. This later is used in the cost function along with orientation_scale as the first component. Frontier orientation is visualised as an arrow placed at frontier centroid position.